Renamed PhotoUploadConfirmation.js to PhotoNotification.js
